Add options for memory- and address sanitizer to CMakeLists.txt
CMakeLists.txt now supports the following options
  - MSAN, boolean, indicating whether to use the memory sanitizer
  - ASAN, boolean, indicating whether to use the address sanitizer

Note that it is not possible to enable both options simultaneously, and
that this is only supported when compiling with clang. Targets using
this library will inherit these options and will also be compiled with
the same sanitizers as enabled for the packet library.